+++ src/main/java/org/amazons/Empty.java
@@ -2,7 +2,7 @@ package org.amazons;
 
 public class Empty extends SquareInfo
 {
-    public Empty(int row, int column) {
+    public Empty(byte row, byte column) {
         this.row = row;
         this.column = column;
     }
